projects/rec-engine/pyproject.toml
[tool.poetry]
name = "rec-engine"
version = "0.1.0"
description = "A Python project by Cheffelo."
authors = ["Mats Eikeland Mollestad <mats.mollestad@cheffelo.com>"]
packages = [{include = "rec_engine"}]

[tool.poetry.dependencies]
python = "^3.10"
pydantic-settings = "^2.1.0"
data-contracts = {path = "../../packages/data-contracts", develop = true, extras = ["mssql"]}
scikit-learn = "^1.4.0"
mlflow = "^2.10.1"
pandera = "^0.18.0"
pydantic-form = {path = "../../packages/pydantic-form", develop = true}
pydantic-argparser = {path = "../../packages/pydantic-argparser", develop = true}
adlfs = "^2024.2.0"

[tool.poetry.group.app.dependencies]
streamlit = "^1.30.0"

[tool.poetry.group.dev.dependencies]
black = "*"
pre-commit = "*"
pytest = "*"
pytest-asyncio = "*"
ruff = "*"
chef = {path = "../../packages/chef", develop = true}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
chef = 'chef.cli:cli'
